refactor(web): extract files/fs mixins from web_server.py (shard s1)
