+++ hermes_cli/web_server.py
@@ -1728,179 +1728,10 @@ _MEDIA_CONTENT_TYPES = {
     ".ico": "image/x-icon",
 }
 _MEDIA_MAX_BYTES = 25 * 1024 * 1024
-_MANAGED_FILES_ROOT_ENV = "HERMES_DASHBOARD_FILES_ROOT"
 _MANAGED_FILE_MAX_BYTES = 100 * 1024 * 1024
-_HOSTED_MANAGED_FILES_ROOT = Path("/opt/data")
-
-
-@dataclass(frozen=True)
-class ManagedFilesPolicy:
-    default_path: Path
-    locked_root: Path | None
-    can_change_path: bool
-
-
-_FS_READDIR_HIDDEN = {
-    ".git",
-    ".hg",
-    ".svn",
-    ".cache",
-    ".next",
-    ".turbo",
-    ".venv",
-    "__pycache__",
-    "build",
-    "dist",
-    "node_modules",
-    "target",
-    "venv",
-}
-
-# Filenames that must never be listed, read, or downloaded through the
-# managed-files API.  These typically contain credentials (API keys, tokens)
-# and exposing them through the dashboard file browser is a security leak —
-# see issue #57505. The set mirrors the credential-file basenames of the two
-# canonical credential guards elsewhere in the codebase
-# (agent.file_safety.get_read_block_error and
-# gateway.platforms.base._ROOT_CREDENTIAL_FILES) so the dashboard Files tab
-# doesn't lag behind them — an operator can point the managed root at
-# HERMES_HOME itself, at which point every one of these basenames is a live
-# secret store sitting in the browsable tree.
-_SENSITIVE_MANAGED_FILE_BASENAMES = frozenset({
-    "auth.json",
-    "auth.lock",
-    "credentials",
-    "config.yaml",
-    ".anthropic_oauth.json",
-    "google_token.json",
-    "google_oauth_pending.json",
-    "google_oauth.json",
-    "webhook_subscriptions.json",
-    "bws_cache.json",
-    "bws_cache.enc.json",
-    # git's credential-store helper cache (agent.file_safety blocks this too).
-    ".git-credentials",
-})
-
-# Directory names whose entire subtree is credential material. Both canonical
-# guards deny these as directory trees, not basenames:
-#   * gateway.platforms.base._ROOT_CREDENTIAL_DIRS = {"pairing", "mcp-tokens"}
-#   * agent.file_safety.get_read_block_error (mcp-tokens/ prefix match)
-# The managed-files API lets the browser descend into subdirs, so a
-# basename-only guard would still expose e.g. ``mcp-tokens/<server>.json``
-# (live MCP OAuth tokens) and ``pairing/<x>``. We match on ANY path component
-# so these trees are blocked wherever they appear under the browsable root,
-# without needing to resolve them relative to HERMES_HOME.
-_SENSITIVE_MANAGED_DIR_NAMES = frozenset({
-    "mcp-tokens",
-    "pairing",
-})
-
-
-def _is_sensitive_filename(name: str) -> bool:
-    """Return True for a basename the managed-files API must never expose.
-
-    Covers ``.env`` / ``.env.<suffix>`` / ``.envrc`` variants plus the
-    canonical Hermes credential-store basenames (see
-    ``_SENSITIVE_MANAGED_FILE_BASENAMES`` above).
-
-    Case-insensitive so ``.ENV`` / ``.Env.local`` / ``Auth.JSON`` on
-    case-insensitive filesystems (macOS/Windows mounts) can't slip past
-    the guard.
-
-    Basename-only: for the directory-tree credential stores
-    (``mcp-tokens/``, ``pairing/``) that the canonical guards also deny,
-    use :func:`_is_sensitive_path`, which the API call sites route through.
-    """
-    lowered = name.lower()
-    if lowered == ".env" or lowered.startswith(".env.") or lowered == ".envrc":
-        return True
-    return lowered in _SENSITIVE_MANAGED_FILE_BASENAMES
-
-
-def _is_sensitive_path(path: Path) -> bool:
-    """Return True for any path the managed-files API must never expose.
-
-    Combines the basename denylist (:func:`_is_sensitive_filename`) with a
-    credential-directory-tree check: a path is sensitive if its own basename
-    is sensitive OR any of its path components is a credential directory
-    (``mcp-tokens`` / ``pairing``). The component match is case-insensitive
-    and needs no HERMES_HOME resolution, so it blocks these trees wherever
-    they sit under the operator-configured managed root — closing the gap
-    the canonical guards cover as directory trees but a basename-only check
-    would miss.
-
-    Read-side only: this guards list/read/download (the #57505 exfil surface).
-    The write endpoints (upload/mkdir/delete) are a separate threat class
-    handled by the write-path checks; extending this guard to them is out of
-    scope for this fix.
-    """
-    if _is_sensitive_filename(path.name):
-        return True
-    return any(part.lower() in _SENSITIVE_MANAGED_DIR_NAMES for part in path.parts)
 
 
 _FS_DATA_URL_MAX_BYTES = 16 * 1024 * 1024
-_FS_TEXT_SOURCE_MAX_BYTES = 64 * 1024 * 1024
-_FS_TEXT_PREVIEW_MAX_BYTES = 512 * 1024
-# Upper bound for the in-app spot editor's save. The editor only opens
-# non-truncated text (<= the preview cap), so this is a safety ceiling against
-# a pasted-in megablob, not the expected payload size.
-_FS_TEXT_WRITE_MAX_BYTES = 8 * 1024 * 1024
-_FS_PREVIEW_LANGUAGE_BY_EXT = {
-    ".c": "c",
-    ".conf": "ini",
-    ".cpp": "cpp",
-    ".css": "css",
-    ".csv": "csv",
-    ".go": "go",
-    ".graphql": "graphql",
-    ".h": "c",
-    ".hpp": "cpp",
-    ".html": "html",
-    ".java": "java",
-    ".js": "javascript",
-    ".json": "json",
-    ".jsx": "jsx",
-    ".kt": "kotlin",
-    ".lua": "lua",
-    ".md": "markdown",
-    ".mjs": "javascript",
-    ".py": "python",
-    ".rb": "ruby",
-    ".rs": "rust",
-    ".sh": "shell",
-    ".sql": "sql",
-    ".svg": "xml",
-    ".toml": "toml",
-    ".ts": "typescript",
-    ".tsx": "tsx",
-    ".txt": "text",
-    ".xml": "xml",
-    ".yaml": "yaml",
-    ".yml": "yaml",
-    ".zsh": "shell",
-}
-_FS_MIME_TYPES = {
-    ".avi": "video/x-msvideo",
-    ".bmp": "image/bmp",
-    ".flac": "audio/flac",
-    ".gif": "image/gif",
-    ".jpeg": "image/jpeg",
-    ".jpg": "image/jpeg",
-    ".m4a": "audio/mp4",
-    ".mkv": "video/x-matroska",
-    ".mov": "video/quicktime",
-    ".mp3": "audio/mpeg",
-    ".mp4": "video/mp4",
-    ".ogg": "audio/ogg",
-    ".opus": "audio/ogg; codecs=opus",
-    ".png": "image/png",
-    ".svg": "image/svg+xml",
-    ".wav": "audio/wav",
-    ".webm": "video/webm",
-    ".webp": "image/webp",
-}
 
 
 def _fs_path(raw_path: str) -> Path:
@@ -1921,89 +1752,6 @@ def _fs_path(raw_path: str) -> Path:
         return candidate.resolve(strict=False)
     except (OSError, RuntimeError, ValueError):
         raise HTTPException(status_code=400, detail="Invalid path")
-
-
-def _fs_mime_type(path: Path) -> str:
-    suffix = path.suffix.lower()
-    if suffix in _FS_MIME_TYPES:
-        return _FS_MIME_TYPES[suffix]
-    guessed, _ = mimetypes.guess_type(str(path))
-    return guessed or "application/octet-stream"
-
-
-def _fs_looks_binary(data: bytes) -> bool:
-    if not data:
-        return False
-    if b"\0" in data:
-        return True
-    suspicious = sum(1 for byte in data if byte < 32 and byte not in {9, 10, 13})
-    return suspicious / len(data) > 0.12
-
-
-def _fs_regular_file(path: Path) -> tuple[Path, os.stat_result]:
-    target = _fs_path(str(path))
-    try:
-        st = target.stat()
-    except FileNotFoundError:
-        raise HTTPException(status_code=404, detail="File not found")
-    except NotADirectoryError:
-        raise HTTPException(status_code=404, detail="File not found")
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not readable")
-    except OSError as exc:
-        raise HTTPException(status_code=400, detail=str(exc) or "Invalid path")
-    if stat.S_ISDIR(st.st_mode):
-        raise HTTPException(status_code=400, detail="Path points to a directory")
-    if not stat.S_ISREG(st.st_mode):
-        raise HTTPException(status_code=400, detail="Only regular files can be read")
-    return target, st
-
-
-def _fs_find_git_root(start: Path) -> str | None:
-    directory = start
-    for _ in range(50):
-        try:
-            if (directory / ".git").exists():
-                return str(directory)
-        except OSError:
-            return None
-        parent = directory.parent
-        if parent == directory:
-            return None
-        directory = parent
-    return None
-
-
-def _fs_default_cwd() -> str:
-    cfg_terminal = load_config().get("terminal") or {}
-    raw = str(cfg_terminal.get("cwd") or os.environ.get("TERMINAL_CWD") or "").strip()
-    if raw and raw not in {".", "auto", "cwd"}:
-        try:
-            candidate = Path(raw).expanduser().resolve(strict=False)
-            if candidate.is_dir():
-                return str(candidate)
-        except (OSError, RuntimeError):
-            pass
-    return str(Path.cwd())
-
-
-def _fs_git_branch(cwd: str) -> str:
-    try:
-        run_kwargs: Dict[str, Any] = {
-            "capture_output": True,
-            "text": True,
-            "timeout": 2,
-            "check": False,
-        }
-        if sys.platform == "win32":
-            run_kwargs["creationflags"] = windows_hide_flags()
-        result = subprocess.run(
-            ["git", "-C", cwd, "branch", "--show-current"],
-            **run_kwargs,
-        )
-        return result.stdout.strip() if result.returncode == 0 else ""
-    except Exception:
-        return ""
 
 
 def _media_serve_roots() -> list[Path]:
@@ -2059,47 +1807,8 @@ async def get_media(path: str):
     return {"data_url": f"data:{_MEDIA_CONTENT_TYPES[target.suffix.lower()]};base64,{encoded}"}
 
 
-def _canonical_path(path: Path, *, require_exists: bool = False) -> Path:
-    try:
-        return path.expanduser().resolve(strict=require_exists)
-    except FileNotFoundError:
-        if require_exists:
-            raise HTTPException(status_code=404, detail="Path not found")
-        raise
-    except (OSError, RuntimeError):
-        raise HTTPException(status_code=400, detail="Invalid path")
-
-
-def _ensure_managed_root(raw_path: str | Path) -> Path:
-    root = Path(raw_path).expanduser()
-    try:
-        root.mkdir(parents=True, exist_ok=True)
-        resolved = root.resolve()
-    except (OSError, RuntimeError) as exc:
-        raise HTTPException(status_code=500, detail=f"Managed files root is unavailable: {exc}")
-    if not resolved.is_dir():
-        raise HTTPException(status_code=500, detail="Managed files root is not a directory")
-    return resolved
-
-
 def _path_is_under(root: Path, target: Path) -> bool:
     return target == root or root in target.parents
-
-
-def _path_text(raw_path: str | None) -> str:
-    text = str(raw_path or "").strip()
-    if "\x00" in text:
-        raise HTTPException(status_code=400, detail="Invalid path")
-    return text
-
-
-def _local_dashboard_request(request: Request) -> bool:
-    if getattr(request.app.state, "auth_required", False):
-        return False
-    host = (request.url.hostname or "").lower()
-    client_host = (request.client.host if request.client else "").lower()
-    local_hosts = {"", "localhost", "127.0.0.1", "::1", "testserver", "testclient"}
-    return host in local_hosts or client_host in local_hosts
 
 
 def _default_hermes_root_is_opt_data() -> bool:
@@ -2153,327 +1862,6 @@ def _dashboard_local_update_managed_externally() -> bool:
     return True
 
 
-def _managed_files_policy(request: Request, *, create_root: bool = True) -> ManagedFilesPolicy:
-    raw_forced_root = os.environ.get(_MANAGED_FILES_ROOT_ENV, "").strip()
-    if raw_forced_root:
-        root = _ensure_managed_root(raw_forced_root) if create_root else _canonical_path(Path(raw_forced_root))
-        return ManagedFilesPolicy(default_path=root, locked_root=root, can_change_path=False)
-
-    # Remote/OAuth access does not imply a hosted container. Users can expose a
-    # local dashboard through the auth gate (for example a macOS launchd install)
-    # and still expect the Files page to browse their local home directory. Lock
-    # to /opt/data only when the installation's Hermes root is actually /opt/data
-    # (the container/hosted layout) or when HERMES_DASHBOARD_FILES_ROOT is set.
-    if _default_hermes_root_is_opt_data():
-        root = _ensure_managed_root(_HOSTED_MANAGED_FILES_ROOT) if create_root else _HOSTED_MANAGED_FILES_ROOT
-        return ManagedFilesPolicy(default_path=root, locked_root=root, can_change_path=False)
-
-    home = _canonical_path(Path.home())
-    return ManagedFilesPolicy(default_path=home, locked_root=None, can_change_path=True)
-
-
-def _resolve_managed_path(
-    raw_path: str | None,
-    request: Request,
-    *,
-    for_write: bool = False,
-) -> tuple[ManagedFilesPolicy, Path, str]:
-    policy = _managed_files_policy(request)
-    text = _path_text(raw_path)
-    root = policy.locked_root
-
-    if root is not None and (not text or text in {".", "/"}):
-        candidate = root
-    elif not text:
-        candidate = policy.default_path
-    else:
-        candidate = Path(text).expanduser()
-        if root is not None and not candidate.is_absolute():
-            if any(part == ".." for part in candidate.parts):
-                raise HTTPException(status_code=400, detail="Path cannot contain '..'")
-            candidate = root / candidate
-        elif not candidate.is_absolute():
-            raise HTTPException(status_code=400, detail="Path must be absolute")
-
-    if ".." in candidate.parts:
-        raise HTTPException(status_code=400, detail="Path cannot contain '..'")
-
-    if for_write and not candidate.exists():
-        parent = _canonical_path(candidate.parent)
-        resolved = parent / candidate.name
-    else:
-        resolved = _canonical_path(candidate, require_exists=not for_write)
-
-    if root is not None and not _path_is_under(root, resolved):
-        raise HTTPException(status_code=403, detail="Path outside managed files root")
-
-    return policy, resolved, str(resolved)
-
-
-def _managed_response_meta(policy: ManagedFilesPolicy) -> Dict[str, Any]:
-    locked_root = str(policy.locked_root) if policy.locked_root is not None else None
-    return {
-        "root": locked_root,
-        "locked_root": locked_root,
-        "can_change_path": policy.can_change_path,
-    }
-
-
-def _managed_file_entry(policy: ManagedFilesPolicy, target: Path) -> Dict[str, Any]:
-    try:
-        resolved = target.resolve()
-    except (OSError, RuntimeError):
-        raise HTTPException(status_code=400, detail="Invalid path")
-    if policy.locked_root is not None and not _path_is_under(policy.locked_root, resolved):
-        raise HTTPException(status_code=403, detail="Path outside managed files root")
-
-    try:
-        st = resolved.stat()
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not stat path: {exc}")
-
-    is_dir = resolved.is_dir()
-    mime_type = None if is_dir else (mimetypes.guess_type(resolved.name)[0] or "application/octet-stream")
-    return {
-        "name": target.name or resolved.name or str(resolved),
-        "path": str(resolved),
-        "is_directory": is_dir,
-        "size": None if is_dir else st.st_size,
-        "mtime": st.st_mtime,
-        "mime_type": mime_type,
-    }
-
-
-def _decode_data_url(data_url: str) -> tuple[bytes, str]:
-    text = (data_url or "").strip()
-    if not text.startswith("data:") or "," not in text:
-        raise HTTPException(status_code=400, detail="Upload payload must be a data URL")
-    header, encoded = text.split(",", 1)
-    mime_type = header[5:].split(";", 1)[0] or "application/octet-stream"
-    if ";base64" not in header:
-        raise HTTPException(status_code=400, detail="Upload payload must be base64 encoded")
-    try:
-        data = base64.b64decode(encoded, validate=True)
-    except (binascii.Error, ValueError):
-        raise HTTPException(status_code=400, detail="Upload payload is not valid base64")
-    if len(data) > _MANAGED_FILE_MAX_BYTES:
-        raise HTTPException(status_code=413, detail="File is too large")
-    return data, mime_type
-
-
-_CHAT_IMAGE_UPLOAD_MAX_BYTES = 25 * 1024 * 1024
-_CHAT_IMAGE_ALLOWED_EXTENSIONS = frozenset({".png", ".jpg", ".jpeg", ".gif", ".webp", ".bmp"})
-_CHAT_IMAGE_MAGIC: tuple[tuple[bytes, str], ...] = (
-    (b"\x89PNG\r\n\x1a\n", ".png"),
-    (b"\xff\xd8\xff", ".jpg"),
-    (b"GIF87a", ".gif"),
-    (b"GIF89a", ".gif"),
-    (b"BM", ".bmp"),
-)
-
-
-def _sanitize_chat_image_filename(filename: str | None) -> str:
-    candidate = Path(str(filename or "").strip()).name
-    candidate = re.sub(r"[\x00-\x1f]+", "_", candidate)
-    candidate = candidate.strip().strip(".")
-    return candidate or "pasted-image"
-
-
-def _chat_image_extension(data: bytes) -> str | None:
-    head = data[:16]
-    if head.startswith(b"RIFF") and head[8:12] == b"WEBP":
-        return ".webp"
-    for sig, ext in _CHAT_IMAGE_MAGIC:
-        if head.startswith(sig):
-            return ext
-    return None
-
-
-def _decode_chat_image_upload(payload: ChatImageUpload) -> tuple[bytes, str, str]:
-    data, mime_type = _decode_data_url(payload.data_url)
-    if not mime_type.lower().startswith("image/"):
-        raise HTTPException(status_code=400, detail="Upload payload must be an image")
-    if len(data) > _CHAT_IMAGE_UPLOAD_MAX_BYTES:
-        mb = _CHAT_IMAGE_UPLOAD_MAX_BYTES // (1024 * 1024)
-        raise HTTPException(status_code=413, detail=f"Image is too large; cap is {mb} MB")
-
-    ext = _chat_image_extension(data)
-    if ext not in _CHAT_IMAGE_ALLOWED_EXTENSIONS:
-        raise HTTPException(status_code=400, detail="Unsupported image type")
-    return data, mime_type, ext
-
-
-@app.post("/api/chat/image-upload")
-async def upload_chat_image(payload: ChatImageUpload, profile: Optional[str] = None):
-    """Persist a browser-provided chat image where the embedded TUI can read it.
-
-    The dashboard /chat page runs Hermes inside an xterm.js PTY. Browser
-    clipboard image bytes are not visible to the server-side clipboard, so the
-    page uploads them here, then drives the TUI's ``/image <path>`` command
-    with the returned gateway-visible path. Files land under
-    ``HERMES_HOME/images/`` — the same directory ``clipboard.paste`` /
-    ``image.attach`` already use.
-    """
-    data, mime_type, ext = _decode_chat_image_upload(payload)
-    with _profile_scope(profile) as scoped_home:
-        home = scoped_home or get_hermes_home()
-        img_dir = Path(home) / "images"
-        try:
-            img_dir.mkdir(parents=True, exist_ok=True)
-        except PermissionError:
-            raise HTTPException(status_code=403, detail="Image directory is not writable")
-        except OSError as exc:
-            raise HTTPException(status_code=500, detail=f"Could not create image directory: {exc}")
-
-        stem = Path(_sanitize_chat_image_filename(payload.filename)).stem or "pasted-image"
-        stem = re.sub(r"[^A-Za-z0-9_.-]+", "_", stem).strip("._-") or "pasted-image"
-        ts = datetime.now().strftime("%Y%m%d_%H%M%S")
-        target = img_dir / f"dashboard_{ts}_{secrets.token_hex(4)}_{stem}{ext}"
-
-        try:
-            target.write_bytes(data)
-        except PermissionError:
-            raise HTTPException(status_code=403, detail="Image directory is not writable")
-        except OSError as exc:
-            raise HTTPException(status_code=500, detail=f"Could not write image: {exc}")
-
-    return {
-        "ok": True,
-        "path": str(target),
-        "name": target.name,
-        "bytes": len(data),
-        "mime_type": mime_type,
-    }
-
-
-@app.get("/api/files")
-async def list_managed_files(request: Request, path: Optional[str] = None):
-    policy, target, display_path = _resolve_managed_path(path, request)
-    if not target.exists():
-        raise HTTPException(status_code=404, detail="Path not found")
-    if not target.is_dir():
-        raise HTTPException(status_code=400, detail="Path is not a directory")
-
-    try:
-        entries = [
-            _managed_file_entry(policy, child)
-            for child in target.iterdir()
-            if not _is_sensitive_path(child)
-        ]
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="Directory is not readable")
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not read directory: {exc}")
-
-    entries.sort(key=lambda item: (not item["is_directory"], str(item["name"]).lower()))
-    locked_root = policy.locked_root
-    parent = None
-    if target.parent != target and (locked_root is None or target != locked_root):
-        parent = str(target.parent)
-    return {
-        "path": display_path,
-        "parent": parent,
-        "entries": entries,
-        **_managed_response_meta(policy),
-    }
-
-
-@app.get("/api/files/read")
-async def read_managed_file(request: Request, path: str):
-    policy, target, display_path = _resolve_managed_path(path, request)
-    if not target.exists():
-        raise HTTPException(status_code=404, detail="File not found")
-    if not target.is_file():
-        raise HTTPException(status_code=400, detail="Path is not a file")
-    if _is_sensitive_path(target):
-        raise HTTPException(status_code=403, detail="Access to sensitive files is not allowed")
-
-    try:
-        size = target.stat().st_size
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not stat file: {exc}")
-    if size > _MANAGED_FILE_MAX_BYTES:
-        raise HTTPException(status_code=413, detail="File is too large")
-
-    mime_type = mimetypes.guess_type(target.name)[0] or "application/octet-stream"
-    try:
-        encoded = base64.b64encode(target.read_bytes()).decode("ascii")
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not readable")
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not read file: {exc}")
-
-    return {
-        "name": target.name,
-        "path": display_path,
-        "size": size,
-        "mime_type": mime_type,
-        "data_url": f"data:{mime_type};base64,{encoded}",
-        **_managed_response_meta(policy),
-    }
-
-
-@app.get("/api/files/download")
-async def download_managed_file(request: Request, path: str):
-    """Stream a managed file as an attachment download.
-
-    Remote clients (desktop app, browser dashboard) open agent-written files
-    that live on *this* gateway's disk, not theirs. Auth-gated like every other
-    managed-files route — ``auth_middleware`` additionally accepts the session
-    token as a ``?token=`` query param here so a shell/browser-opened download
-    (which can't set the session header) still authenticates. See ``/api/pty``
-    for the same query-token precedent.
-    """
-    policy, target, _display_path = _resolve_managed_path(path, request)
-    if not target.exists():
-        raise HTTPException(status_code=404, detail="File not found")
-    if not target.is_file():
-        raise HTTPException(status_code=400, detail="Path is not a file")
-    if _is_sensitive_path(target):
-        raise HTTPException(status_code=403, detail="Access to sensitive files is not allowed")
-
-    try:
-        size = target.stat().st_size
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not stat file: {exc}")
-    if size > _MANAGED_FILE_MAX_BYTES:
-        raise HTTPException(status_code=413, detail="File is too large")
-
-    mime_type = mimetypes.guess_type(target.name)[0] or "application/octet-stream"
-
-    return FileResponse(
-        path=str(target),
-        media_type=mime_type,
-        filename=target.name,
-        content_disposition_type="attachment",
-    )
-
-
-@app.post("/api/files/upload")
-async def upload_managed_file(payload: ManagedFileUpload, request: Request):
-    policy, target, display_path = _resolve_managed_path(payload.path, request, for_write=True)
-    if target.exists() and target.is_dir():
-        raise HTTPException(status_code=409, detail="A directory already exists at that path")
-    if target.exists() and not payload.overwrite:
-        raise HTTPException(status_code=409, detail="File already exists")
-
-    data, _mime_type = _decode_data_url(payload.data_url)
-    try:
-        target.parent.mkdir(parents=True, exist_ok=True)
-        target.write_bytes(data)
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not writable")
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not write file: {exc}")
-
-    return {
-        "ok": True,
-        "entry": _managed_file_entry(policy, target),
-        "path": display_path,
-        **_managed_response_meta(policy),
-    }
-
-
 # Stream uploads to disk in fixed-size chunks. The legacy JSON endpoint above
 # buffers the whole file as a base64 data URL in a JSON body, which (a) inflates
 # the payload ~33%, (b) holds the entire file (plus its decoded copy) in memory,
@@ -2484,242 +1872,65 @@ async def upload_managed_file(payload: ManagedFileUpload, request: Request):
 _UPLOAD_CHUNK_BYTES = 1024 * 1024
 
 
-@app.post("/api/files/upload-stream")
-async def upload_managed_file_stream(
-    request: Request,
-    file: UploadFile = File(...),
-    path: str = Form(...),
-    overwrite: bool = Form(True),
-):
-    policy, target, display_path = _resolve_managed_path(path, request, for_write=True)
-    if target.exists() and target.is_dir():
-        raise HTTPException(status_code=409, detail="A directory already exists at that path")
-    if target.exists() and not overwrite:
-        raise HTTPException(status_code=409, detail="File already exists")
+from hermes_cli.web_routers import files as _files_routes  # noqa: E402
 
-    try:
-        target.parent.mkdir(parents=True, exist_ok=True)
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not writable")
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not create parent directory: {exc}")
+app.include_router(_files_routes.router)
+from hermes_cli.web_routers.files import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    ManagedFilesPolicy,
+    _MANAGED_FILES_ROOT_ENV,
+    _HOSTED_MANAGED_FILES_ROOT,
+    _SENSITIVE_MANAGED_FILE_BASENAMES,
+    _SENSITIVE_MANAGED_DIR_NAMES,
+    _CHAT_IMAGE_UPLOAD_MAX_BYTES,
+    _CHAT_IMAGE_ALLOWED_EXTENSIONS,
+    _CHAT_IMAGE_MAGIC,
+    _is_sensitive_filename,
+    _is_sensitive_path,
+    _canonical_path,
+    _ensure_managed_root,
+    _path_text,
+    _local_dashboard_request,
+    _managed_files_policy,
+    _resolve_managed_path,
+    _managed_response_meta,
+    _managed_file_entry,
+    _decode_data_url,
+    _sanitize_chat_image_filename,
+    _chat_image_extension,
+    _decode_chat_image_upload,
+    upload_chat_image,
+    list_managed_files,
+    read_managed_file,
+    download_managed_file,
+    upload_managed_file,
+    upload_managed_file_stream,
+    create_managed_directory,
+    delete_managed_file,
+)
 
-    # Write to a sibling temp file first so a partial/aborted upload never
-    # clobbers an existing file, then atomically rename into place.
-    tmp_fd, tmp_name = tempfile.mkstemp(
-        prefix=f".{target.name}.", suffix=".upload", dir=str(target.parent)
-    )
-    tmp_path = Path(tmp_name)
-    total = 0
-    renamed = False
-    try:
-        with os.fdopen(tmp_fd, "wb") as out:
-            while True:
-                chunk = await file.read(_UPLOAD_CHUNK_BYTES)
-                if not chunk:
-                    break
-                total += len(chunk)
-                if total > _MANAGED_FILE_MAX_BYTES:
-                    raise HTTPException(status_code=413, detail="File is too large")
-                out.write(chunk)
-        os.replace(tmp_path, target)
-        renamed = True
-    except HTTPException:
-        raise
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not writable")
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not write file: {exc}")
-    finally:
-        # Clean up the temp file on every non-success exit, including
-        # BaseException paths the `except` clauses above don't catch — most
-        # importantly asyncio.CancelledError when a browser aborts a large
-        # upload mid-stream (the exact NS-501 scenario). os.replace clears
-        # tmp_path on success, so only unlink when the rename didn't happen.
-        if not renamed:
-            tmp_path.unlink(missing_ok=True)
-        await file.close()
+from hermes_cli.web_routers import fs as _fs_routes  # noqa: E402
 
-    return {
-        "ok": True,
-        "entry": _managed_file_entry(policy, target),
-        "path": display_path,
-        **_managed_response_meta(policy),
-    }
-
-
-@app.post("/api/files/mkdir")
-async def create_managed_directory(payload: ManagedDirectoryCreate, request: Request):
-    policy, target, display_path = _resolve_managed_path(payload.path, request, for_write=True)
-    if target.exists() and not target.is_dir():
-        raise HTTPException(status_code=409, detail="A file already exists at that path")
-
-    try:
-        target.mkdir(parents=True, exist_ok=True)
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="Directory is not writable")
-    except OSError as exc:
-        raise HTTPException(status_code=500, detail=f"Could not create directory: {exc}")
-
-    return {
-        "ok": True,
-        "entry": _managed_file_entry(policy, target),
-        "path": display_path,
-        **_managed_response_meta(policy),
-    }
-
-
-@app.delete("/api/files")
-async def delete_managed_file(payload: ManagedFileDelete, request: Request):
-    policy, target, display_path = _resolve_managed_path(payload.path, request)
-    if policy.locked_root is not None and target == policy.locked_root:
-        raise HTTPException(status_code=400, detail="Cannot delete the managed files root")
-    if target.parent == target:
-        raise HTTPException(status_code=400, detail="Cannot delete the filesystem root")
-    if not target.exists():
-        raise HTTPException(status_code=404, detail="Path not found")
-
-    try:
-        if target.is_dir():
-            if payload.recursive:
-                shutil.rmtree(target)
-            else:
-                target.rmdir()
-        else:
-            target.unlink()
-    except OSError as exc:
-        status_code = 409 if target.is_dir() and not payload.recursive else 500
-        raise HTTPException(status_code=status_code, detail=f"Could not delete path: {exc}")
-
-    return {"ok": True, "path": display_path, **_managed_response_meta(policy)}
-
-
-@app.get("/api/fs/list")
-async def fs_list(path: str):
-    target = _fs_path(path)
-    try:
-        entries = []
-        with os.scandir(target) as scan:
-            for entry in scan:
-                if entry.name in _FS_READDIR_HIDDEN:
-                    continue
-                entries.append({
-                    "name": entry.name,
-                    "path": str(target / entry.name),
-                    "isDirectory": entry.is_dir(follow_symlinks=False),
-                })
-        entries.sort(key=lambda item: (not item["isDirectory"], item["name"].lower(), item["name"]))
-        return {"entries": entries}
-    except FileNotFoundError:
-        return {"entries": [], "error": "ENOENT"}
-    except NotADirectoryError:
-        return {"entries": [], "error": "ENOTDIR"}
-    except PermissionError:
-        return {"entries": [], "error": "EACCES"}
-    except OSError as exc:
-        return {"entries": [], "error": getattr(exc, "strerror", None) or "read-error"}
-
-
-@app.get("/api/fs/read-text")
-async def fs_read_text(path: str):
-    target, st = _fs_regular_file(_fs_path(path))
-    if st.st_size > _FS_TEXT_SOURCE_MAX_BYTES:
-        raise HTTPException(status_code=413, detail="File too large")
-    bytes_to_read = min(st.st_size, _FS_TEXT_PREVIEW_MAX_BYTES)
-    try:
-        with target.open("rb") as handle:
-            data = handle.read(bytes_to_read)
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not readable")
-    except OSError as exc:
-        raise HTTPException(status_code=400, detail=str(exc) or "File read failed")
-    return {
-        "binary": _fs_looks_binary(data[:4096]),
-        "byteSize": st.st_size,
-        "language": _FS_PREVIEW_LANGUAGE_BY_EXT.get(target.suffix.lower(), "text"),
-        "mimeType": _fs_mime_type(target),
-        "path": str(target),
-        "text": data.decode("utf-8", errors="replace"),
-        "truncated": st.st_size > _FS_TEXT_PREVIEW_MAX_BYTES,
-    }
-
-
-@app.post("/api/fs/write-text")
-async def fs_write_text(payload: FsWriteText):
-    """Overwrite (or create) a UTF-8 text file for the in-app spot editor.
-
-    Mirrors the local Electron ``hermes:fs:writeText`` hardening: the path is
-    resolved + validated by ``_fs_path``, the parent directory must already
-    exist (we never build directory trees), only regular files may be replaced,
-    and the payload is size-capped. The write is staged to a sibling temp file
-    and ``os.replace``-d into place so a crash mid-write can't truncate the
-    original. Stale-on-disk detection is the client's job (re-read before save),
-    so both transports behave identically.
-    """
-    target = _fs_path(payload.path)
-    text = payload.content or ""
-    if len(text.encode("utf-8")) > _FS_TEXT_WRITE_MAX_BYTES:
-        raise HTTPException(status_code=413, detail="Content too large")
-
-    try:
-        st: Optional[os.stat_result] = target.stat()
-    except FileNotFoundError:
-        st = None
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not writable")
-    except OSError as exc:
-        raise HTTPException(status_code=400, detail=str(exc) or "Invalid path")
-
-    if st is not None and stat.S_ISDIR(st.st_mode):
-        raise HTTPException(status_code=400, detail="Path points to a directory")
-    if st is not None and not stat.S_ISREG(st.st_mode):
-        raise HTTPException(status_code=400, detail="Only regular files can be written")
-    if not target.parent.is_dir():
-        raise HTTPException(status_code=400, detail="Parent directory does not exist")
-
-    tmp = target.with_name(f".{target.name}.hermes-tmp-{os.getpid()}")
-    try:
-        tmp.write_text(text, encoding="utf-8")
-        os.replace(tmp, target)
-    except PermissionError:
-        tmp.unlink(missing_ok=True)
-        raise HTTPException(status_code=403, detail="File is not writable")
-    except OSError as exc:
-        tmp.unlink(missing_ok=True)
-        raise HTTPException(status_code=500, detail=f"Could not write file: {exc}")
-
-    return {"ok": True, "path": str(target), "byteSize": len(text.encode("utf-8"))}
-
-
-@app.get("/api/fs/read-data-url")
-async def fs_read_data_url(path: str):
-    target, st = _fs_regular_file(_fs_path(path))
-    if st.st_size > _FS_DATA_URL_MAX_BYTES:
-        raise HTTPException(status_code=413, detail="File too large")
-    try:
-        encoded = base64.b64encode(target.read_bytes()).decode("ascii")
-    except PermissionError:
-        raise HTTPException(status_code=403, detail="File is not readable")
-    except OSError as exc:
-        raise HTTPException(status_code=400, detail=str(exc) or "File read failed")
-    return {"dataUrl": f"data:{_fs_mime_type(target)};base64,{encoded}"}
-
-
-@app.get("/api/fs/git-root")
-async def fs_git_root(path: str):
-    target = _fs_path(path)
-    try:
-        st = target.stat()
-        start = target if stat.S_ISDIR(st.st_mode) else target.parent
-    except OSError:
-        start = target
-    return {"root": _fs_find_git_root(start)}
-
-
-@app.get("/api/fs/default-cwd")
-async def fs_default_cwd():
-    cwd = _fs_default_cwd()
-    return {"cwd": cwd, "branch": _fs_git_branch(cwd)}
+app.include_router(_fs_routes.router)
+from hermes_cli.web_routers.fs import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
+    _FS_READDIR_HIDDEN,
+    _FS_TEXT_SOURCE_MAX_BYTES,
+    _FS_TEXT_PREVIEW_MAX_BYTES,
+    _FS_TEXT_WRITE_MAX_BYTES,
+    _FS_PREVIEW_LANGUAGE_BY_EXT,
+    _FS_MIME_TYPES,
+    _fs_mime_type,
+    _fs_looks_binary,
+    _fs_regular_file,
+    _fs_find_git_root,
+    _fs_default_cwd,
+    _fs_git_branch,
+    fs_list,
+    fs_read_text,
+    fs_write_text,
+    fs_read_data_url,
+    fs_git_root,
+    fs_default_cwd,
+)
 
 
 # ---------------------------------------------------------------------------
@@ -2771,42 +1982,6 @@ from hermes_cli.web_routers.git import (  # noqa: E402,F401 — legacy re-export
     git_worktree_remove_route,
     git_branch_switch_route,
 )
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 # Host TCP ports each port-binding gateway platform listens on, as
@@ -4807,8 +3982,6 @@ from hermes_cli.web_routers.profiles import (  # noqa: E402,F401 — legacy re-e
 )
 
 
-
-
 app.include_router(_sessions_routes.search_router)
 from hermes_cli.web_routers.sessions import (  # noqa: E402,F401 — legacy re-exports; tests call these via web_server.<name>
     search_sessions,
@@ -6747,8 +5920,6 @@ def _apply_model_assignment_sync(
     }
 
 
-
-
 def _infer_provider_on_model_change(model_val: str, prev_provider: str) -> tuple[str, str]:
     """Infer which provider serves ``model_val`` when the flat Config-page Model
     field changes, given the previously-saved ``prev_provider``.
@@ -8397,8 +7568,6 @@ def _messaging_platform_payload(
 
 def _write_platform_enabled(platform_id: str, enabled: bool) -> None:
     write_platform_config_field(platform_id, "enabled", enabled)
-
-
 
 
 from hermes_cli.web_routers import whatsapp_onboarding as _whatsapp_routes  # noqa: E402
@@ -10542,7 +9711,6 @@ async def cancel_oauth_session(
 # ---------------------------------------------------------------------------
 
 
-
 def _session_latest_descendant(session_id: str, db):
     """Resolve a session id to the newest child leaf session.
 
@@ -10672,14 +9840,6 @@ from hermes_cli.web_routers.sessions import (  # noqa: E402,F401 — legacy re-e
     export_session_endpoint,
     prune_sessions_endpoint,
 )
-
-
-
-
-
-
-
-
 
 
 # Serialises the one-time writable schema bootstrap for read-only opens.
@@ -10825,18 +9985,6 @@ async def _auto_archive_ticker_loop(
         await asyncio.sleep(interval_s)
 
 
-
-
-
-
-
-
-
-
-
-
-
-
 def _prune_sessions(body: SessionPrune):
     """Delete ended sessions matching filters (mirrors `hermes sessions prune`)."""
     has_window = (
@@ -10924,8 +10072,6 @@ def _prune_sessions(body: SessionPrune):
         return {"ok": True, "removed": removed}
     finally:
         db.close()
-
-
 
 
 # ---------------------------------------------------------------------------
@@ -11257,8 +10403,6 @@ def _get_cron_job_sync(job_id: str, profile: Optional[str] = None):
     return job
 
 
-
-
 def _list_cron_job_runs_sync(job_id: str, profile: Optional[str] = None, limit: int = 20):
     """Run sessions produced by a cron job, newest first.
 
@@ -11304,8 +10448,6 @@ def _list_cron_job_runs_sync(job_id: str, profile: Optional[str] = None, limit: 
         db.close()
 
 
-
-
 def _create_cron_job_sync(body: CronJobCreate, profile: Optional[str] = None):
     try:
         profile_name, profile_home = _cron_profile_home(profile)
@@ -11344,10 +10486,6 @@ def _create_cron_job_sync(body: CronJobCreate, profile: Optional[str] = None):
         raise HTTPException(status_code=400, detail=str(e))
 
 
-
-
-
-
 def _update_cron_job_sync(job_id: str, body: CronJobUpdate, profile: Optional[str] = None):
     selected = profile or _find_cron_job_profile(job_id)
     if not selected:
@@ -11382,8 +10520,6 @@ def _update_cron_job_sync(job_id: str, body: CronJobUpdate, profile: Optional[st
     return job
 
 
-
-
 def _pause_cron_job_sync(job_id: str, profile: Optional[str] = None):
     selected = profile or _find_cron_job_profile(job_id)
     if not selected:
@@ -11392,8 +10528,6 @@ def _pause_cron_job_sync(job_id: str, profile: Optional[str] = None):
     if not job:
         raise HTTPException(status_code=404, detail="Job not found")
     return job
-
-
 
 
 def _resume_cron_job_sync(job_id: str, profile: Optional[str] = None):
@@ -11406,8 +10540,6 @@ def _resume_cron_job_sync(job_id: str, profile: Optional[str] = None):
     return job
 
 
-
-
 def _trigger_cron_job_sync(job_id: str, profile: Optional[str] = None):
     selected = profile or _find_cron_job_profile(job_id)
     if not selected:
@@ -11416,8 +10548,6 @@ def _trigger_cron_job_sync(job_id: str, profile: Optional[str] = None):
     if not job:
         raise HTTPException(status_code=404, detail="Job not found")
     return job
-
-
 
 
 def _delete_cron_job_sync(job_id: str, profile: Optional[str] = None):
@@ -11431,8 +10561,6 @@ def _delete_cron_job_sync(job_id: str, profile: Optional[str] = None):
     if not removed:
         raise HTTPException(status_code=404, detail="Job not found")
     return {"ok": True}
-
-
 
 
 def _fire_cron_job_for_profile(profile: str, job_id: str) -> bool:
@@ -11461,15 +10589,11 @@ def _fire_cron_job_for_profile(profile: str, job_id: str) -> bool:
         reset_hermes_home_override(token)
 
 
-
-
 # ---------------------------------------------------------------------------
 # Automation Blueprints — parameterized automation blueprints. The dashboard renders the
 # slot schema as a form; submitting instantiates a real cron job via the same
 # create_job path. See cron/blueprint_catalog.py for the single source of truth.
 # ---------------------------------------------------------------------------
-
-
 
 
 # ---------------------------------------------------------------------------
@@ -11604,14 +10728,6 @@ from hermes_cli.web_routers.mcp import (  # noqa: E402,F401 — legacy re-export
 )
 
 
-
-
-
-
-
-
-
-
 _MCP_DASHBOARD_OAUTH_TTL = 15 * 60
 _MAX_PENDING_MCP_OAUTH_FLOWS = 8
 _mcp_oauth_flows: dict[str, "DashboardOAuthFlow"] = {}
@@ -11742,18 +10858,6 @@ def _run_dashboard_mcp_oauth(flow, cfg: dict) -> None:
         flow.mark_error(msg)
     finally:
         flow.mark_worker_done()
-
-
-
-
-
-
-
-
-
-
-
-
 
 
 def _mcp_install_action_name(name: str) -> str:
@@ -12728,10 +11832,6 @@ from hermes_cli.web_routers.skills import (  # noqa: E402,F401 — legacy re-exp
 )
 
 
-
-
-
-
 # Human-readable labels for each hub source id (matches `hermes skills search`
 # provenance).  Keep in sync with create_source_router()'s source list.
 _SKILL_HUB_SOURCE_LABELS = {
@@ -12788,14 +11888,6 @@ def _installed_hub_identifiers(profile: Optional[str] = None) -> dict:
         return out
     except Exception:
         return {}
-
-
-
-
-
-
-
-
 
 
 # ---------------------------------------------------------------------------
@@ -13023,30 +12115,6 @@ from hermes_cli.web_routers.profiles import (  # noqa: E402,F401 — legacy re-e
 )
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 # ---------------------------------------------------------------------------
 # Skills & Tools endpoints
 #
@@ -13166,8 +12234,6 @@ from hermes_cli.web_routers.skills import (  # noqa: E402,F401 — legacy re-exp
 )
 
 
-
-
 def _clear_skills_prompt_cache() -> None:
     """Best-effort: invalidate the skills system-prompt snapshot after a write.
 
@@ -13179,12 +12245,6 @@ def _clear_skills_prompt_cache() -> None:
         clear_skills_system_prompt_cache(clear_snapshot=True)
     except Exception:
         pass
-
-
-
-
-
-
 
 
 from hermes_cli.web_routers import tools as _tools_routes  # noqa: E402
@@ -13204,10 +12264,6 @@ from hermes_cli.web_routers.tools import (  # noqa: E402,F401 — legacy re-expo
     get_computer_use_status,
     grant_computer_use_permissions,
 )
-
-
-
-
 
 
 # Toolsets whose backends carry a selectable model catalog, mapped to the
@@ -13265,16 +12321,6 @@ def _find_toolset_provider_row(ts_key: str, config: dict, provider: Optional[str
     return next(
         (p for p in rows if _is_provider_active(p, config, force_fresh=True)), None
     )
-
-
-
-
-
-
-
-
-
-
 
 
 # ---------------------------------------------------------------------------
@@ -13442,10 +12488,6 @@ def _probe_terminal_backend(name: str, terminal_cfg: dict) -> tuple:
         return ("unavailable", f"Probe failed: {exc}")
 
 
-
-
-
-
 # ---------------------------------------------------------------------------
 # Computer Use (cua-driver) — cross-platform readiness + macOS permission grant
 #
@@ -13456,10 +12498,6 @@ def _probe_terminal_backend(name: str, terminal_cfg: dict) -> tuple:
 # `cua-driver doctor`. The grant flow is macOS-only (no TCC toggles to request
 # on Windows/Linux).
 # ---------------------------------------------------------------------------
-
-
-
-
 
 
 # ---------------------------------------------------------------------------
